# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ToBu-Finance API (v2 schema)
  version: 1.0.1
  description: |
    REST API for ToBu-Finance using the v2 data model.
    Conventions:
      - ObjectIds are 24-hex strings
      - Money values are integer cents
      - Month inputs are strings in YYYY-MM (server stores month as Date at UTC 1st day)
servers:
  - url: http://localhost:4000
    description: Local dev
paths:
...
  /api/members:
    get:
      summary: List members
      parameters:
        - in: query
          name: userId
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
    post:
      summary: Create member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
  /api/members/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete member
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/accounts:
    get:
      summary: List accounts
      parameters:
        - in: query
          name: memberId
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
    post:
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
  /api/accounts/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/accounts/{id}/members:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      summary: Add or update a member's role on the account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountMemberAdd' }
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
  /api/accounts/{id}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
      - name: memberId
        in: path
        required: true
        schema: { $ref: '#/components/schemas/ObjectId' }
    delete:
      summary: Remove member from account
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/categories:
    get:
      summary: List categories
      parameters:
        - in: query
          name: accountId
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
  /api/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete category
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/member-incomes:
    get:
      summary: List member incomes
      parameters:
        - in: query
          name: accountId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: memberId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: from
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: to
          schema: { $ref: '#/components/schemas/Month' }
      responses:
        '200':
          description: Incomes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MemberIncome' }
    post:
      summary: Create member income
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberIncomeCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberIncome' }
  /api/member-incomes/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update member income
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberIncomeUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberIncome' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete member income
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/category-budgets:
    get:
      summary: List category budgets
      parameters:
        - in: query
          name: accountId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: categoryId
          schema: { $ref: '#/components/schemas/ObjectId' }
        - in: query
          name: from
          schema: { $ref: '#/components/schemas/Month' }
        - in: query
          name: to
          schema: { $ref: '#/components/schemas/Month' }
      responses:
        '200':
          description: Budgets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoryBudget' }
    post:
      summary: Create category budget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryBudgetCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryBudget' }
  /api/category-budgets/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      summary: Update category budget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryBudgetUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryBudget' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete category budget
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: 'NOT_FOUND' }
  schemas:
    ObjectId:
      type: string
      pattern: "^[a-fA-F\\d]{24}$"
      example: "60d21b4667d0d8992e610c85"
    Month:
      type: string
      pattern: "^\\d{4}-(0[1-9]|1[0-2])$"
      example: "2025-08"
    Split:
      type: object
      required: [memberId, split]
      properties:
        memberId: { $ref: '#/components/schemas/ObjectId' }
        split:
          type: number
          minimum: 0
          maximum: 100
    Member:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        userId: { $ref: '#/components/schemas/ObjectId', nullable: true }
    MemberCreate:
      type: object
      properties:
        name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        userId: { $ref: '#/components/schemas/ObjectId', nullable: true }
    MemberUpdate:
      allOf: [ { $ref: '#/components/schemas/MemberCreate' } ]
    AccountMember:
      type: object
      required: [memberId]
      properties:
        memberId: { $ref: '#/components/schemas/ObjectId' }
        role: { type: string, enum: [owner, member], default: member }
    Account:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        name: { type: string, nullable: true }
        currency: { type: string, example: EUR }
        members:
          type: array
          items: { $ref: '#/components/schemas/AccountMember' }
        settings:
          type: object
          properties:
            monthGranularity: { type: string, example: "YYYY-MM" }
    AccountCreate:
      type: object
      properties:
        name: { type: string, nullable: true }
        currency: { type: string, example: EUR, default: EUR }
        members:
          type: array
          items: { $ref: '#/components/schemas/AccountMember' }
        settings:
          type: object
          properties:
            monthGranularity: { type: string, example: "YYYY-MM" }
    AccountUpdate:
      allOf: [ { $ref: '#/components/schemas/AccountCreate' } ]
    AccountMemberAdd:
      type: object
      required: [memberId]
      properties:
        memberId: { $ref: '#/components/schemas/ObjectId' }
        role: { type: string, enum: [owner, member], default: member }
    Category:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        accountId: { $ref: '#/components/schemas/ObjectId' }
        name: { type: string, nullable: true }
        customSplit:
          type: array
          items: { $ref: '#/components/schemas/Split' }
    CategoryCreate:
      type: object
      required: [accountId]
      properties:
        accountId: { $ref: '#/components/schemas/ObjectId' }
        name: { type: string, nullable: true }
        customSplit:
          type: array
          items: { $ref: '#/components/schemas/Split' }
    CategoryUpdate:
      allOf: [ { $ref: '#/components/schemas/CategoryCreate' } ]
...
    MemberIncome:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        accountId: { $ref: '#/components/schemas/ObjectId' }
        memberId: { $ref: '#/components/schemas/ObjectId' }
        amountCents: { type: integer, minimum: 0 }
        fromMonth: { type: string, format: date-time, nullable: true }
        toMonth: { type: string, format: date-time, nullable: true }
        source: { type: string, nullable: true }
        note: { type: string, nullable: true }
    MemberIncomeCreate:
      type: object
      required: [accountId, memberId, amountCents]
      properties:
        accountId: { $ref: '#/components/schemas/ObjectId' }
        memberId: { $ref: '#/components/schemas/ObjectId' }
        amountCents: { type: integer, minimum: 0 }
        fromMonth: { $ref: '#/components/schemas/Month' }
        toMonth: { $ref: '#/components/schemas/Month' }
        source: { type: string, nullable: true }
        note: { type: string, nullable: true }
    MemberIncomeUpdate:
      allOf: [ { $ref: '#/components/schemas/MemberIncomeCreate' } ]
...